% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\docType{data}
\name{default_formatters}
\alias{default_formatters}
\title{A list of default formatter mappings}
\usage{
default_formatters
}
\description{
This list matches the most normal mime types with their respective formatters
using default arguments. For a no-frills request parsing this can be supplied
directly to \code{Response$format()}. To add or modify to this list simply supply
the additional parsers as second, third, etc, argument and they will
overwrite or add depending on whether it specifies a mime type already
present.
}
\examples{
\dontrun{
res$format(default_formatters, 'text/plain' = format_plain(sep = ' '))
}

}
\seealso{
\link{formatters} for an overview of the build in formatters in \code{reqres}
}
\keyword{datasets}
