% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\docType{data}
\name{default_parsers}
\alias{default_parsers}
\title{A list of default parser mappings}
\usage{
default_parsers
}
\description{
This list matches the most normal mime types with their respective parsers
using default arguments. For a no-frills request parsing this can be supplied
directly to \code{Request$parse()}. To add or modify to this list simply supply
the additional parsers as second, third, etc, argument and they will
overwrite or add depending on whether it specifies a mime type already
present.
}
\examples{
\dontrun{
req$parse(default_parsers, 'application/json' = parse_json(flatten = TRUE))
}

}
\keyword{datasets}
