% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{maybe_request}
\alias{maybe_request}
\alias{unclass_request}
\title{Check if object is request-like}
\usage{
maybe_request(x)

unclass_request(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
A boolean indicating with some certainty if the object is a request
}
\description{
This function checks an object without looking at the class. Due to a slight
overhead in R6 method dispatch there can be a small gain by removing the
class of the object and treating it as a bare environment. However, in that
case \link{is.Request} will no longer work
}
\keyword{internal}
