% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqres-package.R
\docType{package}
\name{reqres-package}
\alias{reqres}
\alias{reqres-package}
\title{reqres: Powerful Classes for HTTP Requests and Responses}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

In order to facilitate parsing of http requests and creating 
    appropriate responses this package provides two classes to handle a lot of
    the housekeeping involved in working with http exchanges. The infrastructure
    builds upon the 'rook' specification and is thus well suited to be combined
    with 'httpuv' based web servers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/reqres#reqres}
  \item Report bugs at \url{https://github.com/thomasp85/reqres/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com}

}
